+++ src/Phan/LanguageServer/DefinitionResolver.php
@@ -121,10 +121,7 @@ class DefinitionResolver
             if ($type->isNativeType()) {
                 continue;
             }
-            $class_fqsen = $type->asFQSEN();
-            if (!$class_fqsen instanceof FullyQualifiedClassName) {
-                continue;
-            }
+            $class_fqsen = FullyQualifiedClassName::fromType($type);
             if (!$code_base->hasClassWithFQSEN($class_fqsen)) {
                 continue;
             }